Basic styling in place for login modal
I have added some js to open the modal when clicking the login button
in header. This is currently in the wrong file, I will move it later.

Modal also currently can't be closed, I am planning to add a small X to
the top right of the login modal to close it, we should also allow it to
be closed when clicking anywhere outside of the modal itself.