Added ajax request to handle login form with outline of structure for error handling.
Moved some code for loginModal out of modalHelpers into its own file as
it was starting to get a bit long.

I am not sure what the best way to handle ajax request failures, at the
moment I am just alerting the user that there was an error and to try
again.