Save cards to localstorage when loading from server.
This has the issue of overwriting any currently stored cards in
localstorage so we are assuming that the server is the source of truth
for all tests.

I would like to be able to store a test in localstorage and load a
different test from the server at the same time. I would actually like
to be able to store multiple users results in the same browser at the
same time. This would probably be implemented by storing results
alongside the userID