Completely rewritten how test results are stored
Removed the need for global cards or wordlist variables

Instead test state and wordlist are passed to all functions which
require as variabables.

This commit breaks all links with the server and doesn't work with the
current API version. Need to update code to pull results from server and
to update server results